Change log: -----------
Changed the controls for restarting the level. It was F and Control, now it is
Backspace and Delete. Now it should be much harder to restart the level by accident.

Removed the ability to increase gravity for the fire character when pushing down
using WASD or the arrow keys. May add this later, but I think it was causing a bug
where the fire character could get stuck in the ground, and it was a unused feature.

Fixed typos in the labels in the various levels, and changed the labels to represent
the change in key bindings. Also added a note about jumping by pushing the up arrow or the W key.
(Not a new feature, I just forgot to mention it in the game)

Moved the vignette shader to its own file. Changed the blur strength to a uniform so
later there is the potential to add blur controls in game.

Fixed a bug where the background music player was printing the warning message because
it was in the wrong place. Now it has been moved to the proper place.

Resized all of the labels in the fire (2D) dimension so they have the same sizing.
Moved some of the labels so they are a little easier to read.

Redid the crystal position track in the AnimationPlayer in level 4.1, hopefully fixing the game freezing for some
players. I also slowed down the animation a little bit so it is slightly easier to get through the level.