Change log:
Updated the version label in the main menu.

Moved a couple of the labels in the fire dimension again, now
they are placed where they can be read.

Removed all of the shadows from the directional lights from all of the 3D levels.
The reason behind this is they were not being used anyway since the viewport had
a shadow atlas size of zero. I changed the shadow atlas size to 2048, but the shadows
did not look that good, and I currently do not have the time to tweak them (and that would
probably be against the rules, since the voting period is still going).

I think that is all of the changes. They may have been some really minor changes, but if there are, I don't
remember what those changes were.